Revert preadv off_t changes; keep only deprecated-copy fixes in headers

+++ common/stream.h
@@ -36,7 +36,7 @@ public:
     // should keep read/readv write/writev as photon-safe atomic operation
     virtual ssize_t read(void *buf, size_t count) = 0;
     virtual ssize_t readv(const struct iovec *iov, int iovcnt) = 0;
-    virtual ssize_t preadv(const struct iovec *iov, int iovcnt, off_t off) {
+    virtual ssize_t preadv(const struct iovec *iov, int iovcnt, size_t off) {
         return -ENOTSUP;
     }
     virtual ssize_t readv_mutable(struct iovec *iov, int iovcnt)
